Former-commit-id: a2021f8f883a2162768d8952efb579c52c217503
Former-commit-id: 4a7275be03b7ea11065913940c849228c60ba097

+++ plugins/lua/lib/LuaGlue/include/LuaGlue/LuaGlueFunction.h
@@ -45,7 +45,7 @@ class LuaGlueFunction : public LuaGlueFunctionBase
 		int invoke(lua_State *state)
 		{
 			ReturnType ret = applyTuple(g, state, fn_, args);
-			lua_pop(state, Arg_Count_);
+			lua_pop(state, (int)Arg_Count_);
 			stack<_Ret>::put(g, state, ret);
 			return 1;
 		}
